requirements.txt
flask==3.0.0
flask-compress==1.15
flask-login==0.6.3
flask-wtf==1.2.1
email-validator==2.2.0
bcrypt==4.1.2
google-cloud-secret-manager==2.20.0
google-cloud-firestore==2.19.0
